@@ -1,44 +1,49 @@
 package laquay.com.canalestdt;
 
-import android.content.Intent;
+import android.graphics.Bitmap;
+import android.media.MediaPlayer;
 import android.net.Uri;
+import android.os.Build;
 import android.os.Bundle;
-import android.support.v4.content.ContextCompat;
-import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.Toolbar;
 import android.view.View;
 import android.widget.AdapterView;
 import android.widget.ArrayAdapter;
+import android.widget.ImageView;
 import android.widget.ListView;
 import android.widget.TextView;
+import android.widget.Toast;
 
-import com.google.android.exoplayer2.ExoPlayerFactory;
-import com.google.android.exoplayer2.SimpleExoPlayer;
-import com.google.android.exoplayer2.source.MediaSource;
-import com.google.android.exoplayer2.source.hls.HlsMediaSource;
-import com.google.android.exoplayer2.ui.PlayerView;
-import com.google.android.exoplayer2.upstream.DefaultDataSourceFactory;
-import com.google.android.exoplayer2.util.Util;
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.appcompat.widget.Toolbar;
+import androidx.core.content.ContextCompat;
+import androidx.fragment.app.DialogFragment;
+
+import com.android.volley.Response;
+import com.android.volley.VolleyError;
+import com.android.volley.toolbox.ImageRequest;
 
 import java.util.ArrayList;
 
 import laquay.com.canalestdt.component.ChannelList;
+import laquay.com.canalestdt.controller.VolleyController;
 import laquay.com.canalestdt.model.Channel;
 import laquay.com.canalestdt.model.ChannelOptions;
 
 public class DetailChannelActivity extends AppCompatActivity {
     public static final String TAG = DetailChannelActivity.class.getSimpleName();
     public static final String EXTRA_MESSAGE = "laquay.com.canalestdt.CHANNEL_DETAIL";
+    public static final String EXTRA_TYPE = "laquay.com.canalestdt.CHANNEL_TYPE";
+    public static final String TYPE_TV = "TV";
+    public static final String TYPE_RADIO = "RADIO";
     private Channel channel;
-    private SimpleExoPlayer player;
-    private DefaultDataSourceFactory dataSourceFactory;
+    private String typeOfStream;
 
-    private PlayerView channelVideoView;
-    private TextView channelMockTV;
+    private ImageView channelImageIV;
     private TextView channelNameTV;
     private TextView channelURLTV;
     private TextView channelSourceTV;
     private ListView channelSourceLV;
+    private MediaPlayer mediaPlayer; //TODO This should be moved to another Dialog/Fragment
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -49,13 +54,19 @@ public class DetailChannelActivity extends AppCompatActivity {
         setSupportActionBar(toolbar);
         getSupportActionBar().setDisplayHomeAsUpEnabled(true);
         getSupportActionBar().setDisplayShowHomeEnabled(true);
-        getWindow().setStatusBarColor(ContextCompat.getColor(this, R.color.colorPrimaryDark));
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP) {
+            getWindow().setStatusBarColor(ContextCompat.getColor(this, R.color.colorPrimaryDark));
+        }
 
-        Intent intent = getIntent();
-        ChannelList channelList = (ChannelList) intent.getSerializableExtra(EXTRA_MESSAGE);
-        channel = channelList.getChannel();
-
-        getSupportActionBar().setTitle(channelList.getChannel().getName());
+        Bundle intentExtras = getIntent().getExtras();
+        if (intentExtras != null) {
+            ChannelList channelList = (ChannelList) intentExtras.getSerializable(EXTRA_MESSAGE);
+            if (channelList != null) {
+                channel = channelList.getChannel();
+                getSupportActionBar().setTitle(channelList.getChannel().getName());
+            }
+            typeOfStream = intentExtras.getString(EXTRA_TYPE);
+        }
 
         setUpElements();
         setUpListeners();
@@ -64,18 +75,11 @@ public class DetailChannelActivity extends AppCompatActivity {
     }
 
     private void setUpElements() {
-        channelVideoView = findViewById(R.id.channel_video_detail_exoplayer);
-        channelMockTV = findViewById(R.id.channel_mock_detail_tv);
+        channelImageIV = findViewById(R.id.channel_image_detail_tv);
         channelNameTV = findViewById(R.id.channel_name_detail_tv);
         channelURLTV = findViewById(R.id.channel_url_detail_tv);
         channelSourceTV = findViewById(R.id.channel_source_detail_tv);
         channelSourceLV = findViewById(R.id.channel_source_detail_lv);
-
-        player = ExoPlayerFactory.newSimpleInstance(this);
-
-        // Produces DataSource instances through which media data is loaded.
-        dataSourceFactory = new DefaultDataSourceFactory(this,
-                Util.getUserAgent(this, "laquay.com.canalestdt"));
     }
 
     private void setUpListeners() {
@@ -102,36 +106,55 @@ public class DetailChannelActivity extends AppCompatActivity {
                 @Override
                 public void onItemClick(AdapterView<?> parent, View view, int position, long id) {
                     String source = (String) channelSourceLV.getItemAtPosition(position);
-                    loadVideo(source, position);
+                    switch (typeOfStream) {
+                        case TYPE_TV:
+                            loadVideo(source);
+                            break;
+                        case TYPE_RADIO:
+                            loadRadio(source);
+                            break;
+                    }
                 }
             });
-
-            loadVideo(sources.get(0), 0);
-        } else {
-            channelVideoView.setVisibility(View.GONE);
-            channelSourceTV.setVisibility(View.GONE);
-            channelMockTV.setVisibility(View.VISIBLE);
         }
+
+        // Temporary fix
+        String imageUrl = channel.getLogo();
+        imageUrl = imageUrl.replace("http://graph.facebook.com", "https://graph.facebook.com");
+
+        ImageRequest request = new ImageRequest(imageUrl,
+                new Response.Listener<Bitmap>() {
+                    @Override
+                    public void onResponse(Bitmap bitmap) {
+                        channelImageIV.setImageBitmap(bitmap);
+                    }
+                }, 0, 0, ImageView.ScaleType.CENTER_INSIDE, null,
+                new Response.ErrorListener() {
+                    public void onErrorResponse(VolleyError error) {
+                        channelImageIV.setImageResource(R.mipmap.ic_launcher);
+                    }
+                });
+        VolleyController.getInstance(this).addToQueue(request);
     }
 
-    public void loadVideo(String streamURL, int sourceNumber) {
-        channelSourceTV.setText(getString(R.string.channel_detail_currenty_playing) + " - Source: " + (sourceNumber + 1));
-        channelVideoView.setPlayer(player);
-
-        // This is the MediaSource representing the media to be played.
-        MediaSource videoSource = new HlsMediaSource.Factory(dataSourceFactory)
-                .createMediaSource(Uri.parse(streamURL));
-
-        // Prepare the player with the source.
-        player.prepare(videoSource);
+    private void loadVideo(String streamURL) {
+        Toast.makeText(this, getString(R.string.channel_detail_reproducing_tv), Toast.LENGTH_SHORT).show();
+        DialogFragment newFragment = VideoDialogFragment.newInstance(streamURL);
+        newFragment.show(getSupportFragmentManager(), "VideoDialog");
     }
 
-    // Activity onStop, player must be release because of memory saving
+    private void loadRadio(String streamURL) {
+        Toast.makeText(this, getString(R.string.channel_detail_reproducing_radio), Toast.LENGTH_SHORT).show();
+        mediaPlayer = MediaPlayer.create(this, Uri.parse(streamURL));
+        mediaPlayer.start();
+    }
+
     @Override
-    public void onStop() {
-        super.onStop();
-        if (player != null) {
-            player.release();
+    public void onBackPressed() {
+        super.onBackPressed();
+
+        if (mediaPlayer != null) {
+            mediaPlayer.stop();
         }
     }
 
