@@ -1,99 +1,63 @@
 package laquay.com.canalestdt;
 
-import android.content.Intent;
-import android.net.Uri;
 import android.os.Bundle;
-import android.support.design.widget.NavigationView;
-import android.support.v4.app.Fragment;
-import android.support.v4.app.FragmentTransaction;
-import android.support.v4.view.GravityCompat;
-import android.support.v4.widget.DrawerLayout;
-import android.support.v7.app.ActionBarDrawerToggle;
-import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.Toolbar;
 import android.view.MenuItem;
-import android.widget.Toast;
 
-public class MainActivity extends AppCompatActivity
-        implements NavigationView.OnNavigationItemSelectedListener {
-    public static final String TAG = MainActivity.class.getSimpleName();
+import androidx.annotation.NonNull;
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.appcompat.widget.Toolbar;
+import androidx.fragment.app.Fragment;
+import androidx.fragment.app.FragmentTransaction;
+
+import com.google.android.material.bottomnavigation.BottomNavigationView;
+
+import laquay.com.canalestdt.controller.SharedPreferencesController;
+
+public class MainActivity extends AppCompatActivity {
+    private BottomNavigationView.OnNavigationItemSelectedListener mOnNavigationItemSelectedListener
+            = new BottomNavigationView.OnNavigationItemSelectedListener() {
+
+        @Override
+        public boolean onNavigationItemSelected(@NonNull MenuItem item) {
+            int id = item.getItemId();
+
+            Fragment fragment = null;
+            String fragmentTAG = null;
+
+            if (id == R.id.nav_tv_channels) {
+                fragment = TVFragment.newInstance();
+                fragmentTAG = TVFragment.TAG;
+            } else if (id == R.id.nav_radio_channels) {
+                fragment = RadioFragment.newInstance();
+                fragmentTAG = RadioFragment.TAG;
+            } else if (id == R.id.nav_settings) {
+                fragment = SettingsFragment.newInstance();
+                fragmentTAG = SettingsFragment.TAG;
+            }
+
+            if (fragment != null) {
+                FragmentTransaction ft = getSupportFragmentManager().beginTransaction();
+                ft.replace(R.id.content_frame, fragment, fragmentTAG);
+                ft.commit();
+            }
+
+            return true;
+        }
+    };
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
+
         Toolbar toolbar = findViewById(R.id.toolbar);
         setSupportActionBar(toolbar);
 
-        DrawerLayout drawer = findViewById(R.id.drawer_layout);
-        ActionBarDrawerToggle toggle = new ActionBarDrawerToggle(
-                this, drawer, toolbar, R.string.navigation_drawer_open, R.string.navigation_drawer_close);
-        drawer.addDrawerListener(toggle);
-        toggle.syncState();
+        SharedPreferencesController.init(this);
 
-        NavigationView navigationView = findViewById(R.id.nav_view);
-        navigationView.setNavigationItemSelectedListener(this);
+        BottomNavigationView navView = findViewById(R.id.nav_view);
 
-        // Check first item
-        onNavigationItemSelected(navigationView.getMenu().findItem(R.id.nav_tv_channels));
-    }
-
-    @Override
-    public void onBackPressed() {
-        DrawerLayout drawer = findViewById(R.id.drawer_layout);
-        if (drawer.isDrawerOpen(GravityCompat.START)) {
-            drawer.closeDrawer(GravityCompat.START);
-        } else {
-            super.onBackPressed();
-        }
-    }
-
-    @SuppressWarnings("StatementWithEmptyBody")
-    @Override
-    public boolean onNavigationItemSelected(MenuItem item) {
-        int id = item.getItemId();
-
-        Fragment fragment = null;
-        String fragmentTAG = null;
-
-        if (id == R.id.nav_tv_channels) {
-            fragment = MainFragment.newInstance();
-            fragmentTAG = MainFragment.TAG;
-        } else if (id == R.id.nav_radio_channels) {
-            Toast.makeText(this, getString(R.string.toast_function_not_implemented), Toast.LENGTH_SHORT).show();
-        } else if (id == R.id.nav_settings) {
-            Toast.makeText(this, getString(R.string.toast_function_not_implemented), Toast.LENGTH_SHORT).show();
-        } else if (id == R.id.nav_about) {
-            fragment = AboutFragment.newInstance();
-            fragmentTAG = AboutFragment.TAG;
-        } else if (id == R.id.nav_share) {
-            String shareBody = "TDTChannels\n\n" +
-                    "Lista de canales de televisión, y radio, que se emiten en abierto por Internet. " +
-                    "Especialmente enfocado a España, y ampliando a canales Internacionales. Gratuito y sin publicidad.\n\n" +
-                    "https://github.com/LaQuay/TDTChannels";
-            Intent sharingIntent = new Intent(android.content.Intent.ACTION_SEND);
-            sharingIntent.setType("text/plain");
-            sharingIntent.putExtra(android.content.Intent.EXTRA_SUBJECT, "TDT Channels");
-            sharingIntent.putExtra(android.content.Intent.EXTRA_TEXT, shareBody);
-            startActivity(Intent.createChooser(sharingIntent, getString(R.string.share_using)));
-        } else if (id == R.id.nav_send_telegram) {
-            try {
-                Intent intent = new Intent(Intent.ACTION_VIEW, Uri.parse("tg://resolve?domain=canales_tdt"));
-                startActivity(intent);
-            } catch (Exception e) {
-                Intent intent = new Intent(Intent.ACTION_VIEW, Uri.parse("https://t.me/canales_tdt"));
-                startActivity(intent);
-            }
-        }
-
-        if (fragment != null) {
-            FragmentTransaction ft = getSupportFragmentManager().beginTransaction();
-            ft.replace(R.id.content_frame, fragment, fragmentTAG);
-            ft.commit();
-        }
-
-        DrawerLayout drawer = findViewById(R.id.drawer_layout);
-        drawer.closeDrawer(GravityCompat.START);
-        return true;
+        navView.setOnNavigationItemSelectedListener(mOnNavigationItemSelectedListener);
+        navView.setSelectedItemId(R.id.nav_tv_channels);
     }
 }
