@@ -1,5 +1,7 @@
 package laquay.com.canalestdt.utils;
 
 public class APIUtils {
-    public static final String CHANNELS_URL = "http://91.121.64.179/tdt_project/output/channels.json";
+    private static final String BASE_URL = "http://91.121.64.179/tdt_project/output/";
+    public static final String TV_URL = BASE_URL + "channels.json";
+    public static final String RADIO_URL = BASE_URL + "radio_channels.json";
 }
